댓글

+++ SISC/src/main/java/book/controller/BookDeleteController.java
@@ -1,5 +1,7 @@
 package book.controller;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -7,6 +9,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import book.model.BookDao;
+import member.model.MemberBean;
 
 @Controller
 public class BookDeleteController {
@@ -19,8 +22,17 @@ public class BookDeleteController {
 	
 	@RequestMapping(value = command)
 	public String detail(@RequestParam("bk_num") int bk_num, 
-						@RequestParam("pageNumber") int pageNumber, Model model) {
+						@RequestParam("pageNumber") int pageNumber, 
+						Model model, HttpSession session) {
 
+		MemberBean mb = (MemberBean)session.getAttribute(sessionID);
+		
+		if(mb == null) {
+			session.setAttribute("destination", "redirect:detail.bk?bk_num="+bk_num+"&pageNumber="+pageNumber);
+			
+			return "redirect:login.mb";
+		}
+		
 		bdao.deleteBook(bk_num); 
 		
 		model.addAttribute("pageNumber",pageNumber);
